requirements.txt
pyOpenSSL>=23.0.0
requests>=2.31.0
urllib3>=1.26.0,<2.0.0
